+++ src/main/java/org/mmarini/rl/agents/KpiBinWriter.java
@@ -25,14 +25,11 @@
 
 package org.mmarini.rl.agents;
 
-import io.reactivex.rxjava3.core.Completable;
-import io.reactivex.rxjava3.subjects.CompletableSubject;
 import org.nd4j.linalg.api.ndarray.INDArray;
-import org.reactivestreams.Subscriber;
-import org.reactivestreams.Subscription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.Closeable;
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
@@ -48,7 +45,7 @@ import static java.util.Objects.requireNonNull;
  * The data collector consumer accumulates data and returns the kpi of data
  */
 
-public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
+public class KpiBinWriter implements Closeable {
     public static final String[] ANALYSIS_KPIS = {
             "^reward$",
             "^delta$",
@@ -69,7 +66,7 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
             .distinct()
             .toArray(String[]::new);
 
-    private static final Logger logger = LoggerFactory.getLogger(KpiBinSubscriber.class);
+    private static final Logger logger = LoggerFactory.getLogger(KpiBinWriter.class);
 
     /**
      * Returns the subscriber for a given set of keys
@@ -77,13 +74,13 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
      * @param path the files path
      * @param keys the keys
      */
-    public static KpiBinSubscriber create(File path, String... keys) {
+    public static KpiBinWriter create(File path, String... keys) {
         Predicate<String> matchers1 = Arrays.stream(keys)
                 .map(Pattern::compile)
                 .map(Pattern::asPredicate)
                 .reduce(Predicate::or)
                 .orElse(x -> true);
-        return new KpiBinSubscriber(path, matchers1);
+        return new KpiBinWriter(path, matchers1);
     }
 
     /**
@@ -91,8 +88,8 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
      *
      * @param path the files path
      */
-    public static KpiBinSubscriber create(File path) {
-        return new KpiBinSubscriber(path, x -> true);
+    public static KpiBinWriter create(File path) {
+        return new KpiBinWriter(path, x -> true);
     }
 
     /**
@@ -102,11 +99,11 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
      * @param path   the files path
      * @param labels the labels string
      */
-    public static KpiBinSubscriber createFromLabels(File path, String labels) {
+    public static KpiBinWriter createFromLabels(File path, String labels) {
         String[] labs = parseLabels(labels);
         return labs != null
-                ? KpiBinSubscriber.create(path, labs)
-                : KpiBinSubscriber.create(path);
+                ? KpiBinWriter.create(path, labs)
+                : KpiBinWriter.create(path);
     }
 
     /**
@@ -128,44 +125,34 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
 
     private final Predicate<String> matcher;
     private final File path;
-    private final CompletableSubject completed;
     private BinArrayFileMap files;
-    private Subscription subscription;
 
     /**
      * @param path   the path of kpis
      * @param filter the matcher to filter
      */
-    public KpiBinSubscriber(File path, Predicate<String> filter) {
+    public KpiBinWriter(File path, Predicate<String> filter) {
         this.path = requireNonNull(path);
         this.matcher = requireNonNull(filter);
         this.files = BinArrayFileMap.empty();
-        this.completed = CompletableSubject.create();
     }
 
     @Override
-    public void onComplete() {
-        logger.atDebug().log("Complete");
-        try {
-            files.close();
-        } catch (IOException e) {
-            logger.atError().setCause(e).log("Error closing files");
-        }
-        completed.onComplete();
+    public void close() throws IOException {
+        files.close();
+        logger.atInfo().log("Closed kpi files");
     }
 
-    @Override
-    public void onError(Throwable throwable) {
-        logger.atError().setCause(throwable).log(throwable.getMessage());
-        try {
-            files.close();
-        } catch (IOException e) {
-            logger.atError().setCause(e).log("Error closing file");
-        }
+    /**
+     * Flushes the files
+     *
+     * @throws IOException in case of error
+     */
+    public void flush() throws IOException {
+        files.flush();
     }
 
-    @Override
-    public void onNext(Map<String, INDArray> data) {
+    public void write(Map<String, INDArray> data) {
         List<String> keys = data.keySet().stream()
                 .filter(matcher)
                 .toList();
@@ -183,19 +170,5 @@ public class KpiBinSubscriber implements Subscriber<Map<String, INDArray>> {
                 }
             }
         }
-        subscription.request(1);
-    }
-
-    @Override
-    public void onSubscribe(Subscription subscription) {
-        this.subscription = subscription;
-        subscription.request(1);
-    }
-
-    /**
-     * Returns the completion event
-     */
-    public Completable readCompleted() {
-        return completed;
     }
 }