+++ src/main/java/org/mmarini/wheelly/apis/QVect.java
@@ -49,6 +49,20 @@ public record QVect(double[] coords) {
     }
 
     /**
+     * Returns the vector representing the line from the given point to the given direction
+     *
+     * @param point     the point
+     * @param direction the direction
+     */
+    public static QVect line(Point2D point, Complex direction) {
+        double px = point.getX();
+        double py = point.getY();
+        double dx = direction.x();
+        double dy = direction.y();
+        return create(py * dx - px * dy, dy, -dx, 0, 0);
+    }
+
+    /**
      * Returns the zero vector
      */
     public static QVect ones() {
@@ -76,6 +90,35 @@ public record QVect(double[] coords) {
     }
 
     /**
+     * Returns the intersection with a given line (vectors must be a linear vector)
+     *
+     * @param qVect the line
+     */
+    public Point2D intersect(QVect qVect) {
+        double a1 = this.coords[1];
+        double b1 = this.coords[2];
+        double c1 = this.coords[0];
+        double a2 = qVect.coords[1];
+        double b2 = qVect.coords[2];
+        double c2 = qVect.coords[0];
+        double dx = b2 * a1 - b1 * a2;
+        double dy = a2 * b1 - a1 * b2;
+        if (dx == 0 || dy == 0) {
+            return null;
+        }
+        double x = (b1 * c2 - b2 * c1) / dx;
+        double y = (a1 * c2 - a2 * c1) / dy;
+        if (x == -0.0) {
+            x = 0;
+        }
+        if (y == -0.0) {
+            y = 0;
+        }
+
+        return new Point2D.Double(x, y);
+    }
+
+    /**
      * Retruns the scalar product
      *
      * @param other the other QVect
@@ -97,8 +140,7 @@ public record QVect(double[] coords) {
 
     @Override
     public String toString() {
-        String sb = "QVect{" + "coords=" + Arrays.toString(coords) +
+        return "QVect{" + "coords=" + Arrays.toString(coords) +
                 '}';
-        return sb;
     }
 }